Starting level modification loader.
