+++ ab3d2_source/c/level_mod.c
@@ -1,8 +1,11 @@
 #include <stdio.h>
 #include "level_mod.h"
+#include "devmode.h"
+#include <proto/exec.h>
 
+extern char const LMod_PropertiesFile[];
 
-BOOL lmod_ParseLevelMessages(GMF_ChunkHeader const* pChunkHeader, GMF_Data* pGMFData)
+static BOOL lmod_ParseLevelMessages(GMF_ChunkHeader const* pChunkHeader, GMF_Data* pGMFData)
 {
     printf(
         "\tlmod_ParseLevelMessages() %.*s\n",
@@ -28,13 +31,41 @@ static GMF_ParserEntry lmod_Parsers[] = {
     { 0, NULL },
 };
 
-GMF_Data* LMod_LoadFile(char const* filename)
+/**
+ * Level Modification File, LMOD
+ */
+static GMF_Header const lmod_Header = {
+    .h_Ident.id_Value     = IDENT_TKGD,
+    .h_SubFormat.id_Value = IDENT_LMOD,
+    .h_Version            = {TKG_VERSION, TKG_REVISION}
+};
+
+
+/**********************************************************************************************************************/
+
+void LMod_LoadModificationData(void)
 {
-    GMF_Header const lmod_Header = {
-        .h_Ident.id_Value     = IDENT_TKGD,
-        .h_SubFormat.id_Value = IDENT_LMOD,
-        .h_Version            = {TKG_VERSION, TKG_REVISION}
-    };
-    return GMF_LoadFile(filename, &lmod_Header, lmod_Parsers);
+    dputs("LMod_LoadModificationData()");
+    LMod_Properties.lmod_Loaded = GMF_LoadFile(LMod_PropertiesFile, &lmod_Header, lmod_Parsers);
+    if (NULL == LMod_Properties.lmod_Loaded) {
+        LMod_Properties.lmod_PVSErrata = NULL;
+        dputs("No level modification properties loaded");
+        return;
+    }
+    GMF_ChunkHeader const* pChunk = GMF_LocateChunk(LMod_Properties.lmod_Loaded, IDENT_PVSD);
+    LMod_Properties.lmod_PVSErrata = pChunk ? GMF_ChunkData(pChunk) : NULL;
+    dprintf("PVS Errata %p\n", LMod_Properties.lmod_PVSErrata);
+
+    pChunk = GMF_LocateChunk(LMod_Properties.lmod_Loaded, IDENT_BCKD);
+
 }
 
+void LMod_FreeModificationData(void)
+{
+    dputs("LMod_FreeModificationData()");
+    if (LMod_Properties.lmod_Loaded) {
+        FreeVec((void*)LMod_Properties.lmod_Loaded);
+    }
+    LMod_Properties.lmod_Loaded = NULL;
+    LMod_Properties.lmod_PVSErrata = NULL;
+}
