+++ ab3d2_source/macros.i
@@ -406,4 +406,3 @@ DCLC			MACRO
 _\1::
 \1: \2 \3 \4
 				ENDM
-
